refactor(types): tighten boundaries and split provider registry
