+++ gui/tsconfig.node.json
@@ -13,6 +13,7 @@
     "verbatimModuleSyntax": true,
     "moduleDetection": "force",
     "noEmit": true,
+    "strict": true,
 
     /* Linting */
     "noUnusedLocals": true,
